refactor(view): Move logic from BoardCreatorView to controller
